reuse 'main' code in discovery-handler-modules
